refactor: extract chunking functionality into mixin class
