+++ annif/backend/fasttext.py
@@ -7,9 +7,10 @@ from annif.hit import AnalysisHit, ListAnalysisResult
 from annif.exception import NotInitializedException
 import fastText
 from . import backend
+from . import mixins
 
 
-class FastTextBackend(backend.AnnifBackend):
+class FastTextBackend(mixins.ChunkingBackend, backend.AnnifBackend):
     """fastText backend for Annif"""
 
     name = "fasttext"
@@ -125,20 +126,3 @@ class FastTextBackend(backend.AnnifBackend):
                 label=subject[1],
                 score=score / len(chunktexts)))
         return ListAnalysisResult(results, project.subjects)
-
-    def _analyze(self, text, project, params):
-        self.initialize()
-        self.debug('Analyzing text "{}..." (len={})'.format(
-            text[:20], len(text)))
-        sentences = project.analyzer.tokenize_sentences(text)
-        self.debug('Found {} sentences'.format(len(sentences)))
-        chunksize = int(params['chunksize'])
-        chunktexts = []
-        for i in range(0, len(sentences), chunksize):
-            chunktext = ' '.join(sentences[i:i + chunksize])
-            normalized = self._normalize_text(project, chunktext)
-            if normalized != '':
-                chunktexts.append(normalized)
-        self.debug('Split sentences into {} chunks'.format(len(chunktexts)))
-
-        return self._analyze_chunks(chunktexts, project)
